add player class
